Chore: Refactor CLI infrastructure
This refactor cleans up the current code, making easier to add new
features to the CLI. It also homogenizes how the different actions
are tested.
