+++ src/lib/cli/rules/templates/core-rule/rule-doc.hbs
