+++ src/lib/cli/rules/templates/core-rule/rule-script.hbs
