+++ src/lib/cli/rules/templates/core-rule/rule-test.hbs
