+++ tests/lib/cli/rules/create-core-rule.ts
@@ -3,18 +3,18 @@ import * as proxyquire from 'proxyquire';
 import * as sinon from 'sinon';
 import test from 'ava';
 
-import { readFileAsync } from '../../../src/lib/utils/misc';
+import { CLIOptions } from '../../../../src/lib/types';
+import { readFileAsync } from '../../../../src/lib/utils/misc';
 
+const actions = ({ newRule: true } as CLIOptions);
 const ruleScriptDir = 'src/lib/rules';
 const ruleDocDir = 'docs/user-guide/rules';
 const ruleTestDir = 'tests/lib/rules';
-const ruleDistScriptDir = `dist/${ruleScriptDir}`;
 
-const expectedScriptDir = 'tests/lib/cli/fixtures/new.txt';
-const expectedScriptHasQuotesDir = 'tests/lib/cli/fixtures/new-quotes.txt';
-const scriptTemplateDir = 'src/lib/cli/templates/rule-script-ts.hbs';
+const expectedScriptDir = `${__dirname}/fixtures/new.txt`;
+const expectedScriptHasQuotesDir = `${__dirname}/fixtures/new-quotes.txt`;
+const scriptTemplateDir = `src/lib/cli/rules/templates/core-rule/rule-script.hbs`;
 const existingRuleName = 'Content Type';
-const normalizedExistingRuleName = 'content-type';
 const newRuleName = 'new';
 
 const inquirer = { prompt() { } };
@@ -25,7 +25,7 @@ const fs = {
 };
 const rimrafObject = { rimraf() { } };
 const stubPromisifiedMethodObject = {
-    mkdirAsync() { },
+    mkdirPAsync() { },
     readFileAsync() { },
     rimrafAsync() { },
     writeFileAsync() { }
@@ -39,7 +39,7 @@ const stubUtilObject = {
 const resourceLoader = { getCoreRules() { } };
 const rules = ['axe', 'content-type'];
 
-proxyquire('../../../src/lib/cli/rule-generator', {
+proxyquire('../../../../src/lib/cli/rules/create-core-rule', {
     fs,
     inquirer,
     resourceLoader,
@@ -47,17 +47,17 @@ proxyquire('../../../src/lib/cli/rule-generator', {
     util: stubUtilObject
 });
 
-import * as rule from '../../../src/lib/cli/rule-generator';
+import * as rule from '../../../../src/lib/cli/rules/create-core-rule';
 
 test.beforeEach((t) => {
-    sinon.stub(stubPromisifiedMethodObject, 'mkdirAsync').resolves();
+    sinon.stub(stubPromisifiedMethodObject, 'mkdirPAsync').resolves();
     sinon.stub(stubPromisifiedMethodObject, 'rimrafAsync').resolves();
     sinon.stub(stubPromisifiedMethodObject, 'writeFileAsync').resolves();
     sinon.stub(resourceLoader, 'getCoreRules').resolves(rules);
     sinon.spy(stubUtilObject, 'promisify');
 
     t.context.promisify = stubUtilObject.promisify;
-    t.context.mkdirAsync = stubPromisifiedMethodObject.mkdirAsync;
+    t.context.mkdirpAsync = stubPromisifiedMethodObject.mkdirPAsync;
     t.context.rimrafAsync = stubPromisifiedMethodObject.rimrafAsync;
     t.context.writeFileAsync = stubPromisifiedMethodObject.writeFileAsync;
     t.context.loadRules = resourceLoader.getCoreRules;
@@ -65,7 +65,7 @@ test.beforeEach((t) => {
 
 test.afterEach.always((t) => {
     t.context.promisify.restore();
-    t.context.mkdirAsync.restore();
+    t.context.mkdirpAsync.restore();
     t.context.rimrafAsync.restore();
     t.context.writeFileAsync.restore();
     t.context.loadRules.restore();
@@ -86,15 +86,17 @@ test.serial(`if core, 'generate' should call to write script, documentation, tes
     sandbox.stub(stubPromisifiedMethodObject, 'readFileAsync').resolves('');
     sandbox.stub(inquirer, 'prompt').resolves(results);
 
-    await rule.newRule();
+    await rule.newRule(actions);
 
-    const mkdirAsyncFn = t.context.mkdirAsync;
+    const mkdirpAsyncFn = t.context.mkdirpAsync;
     const writeFileAsyncFn = t.context.writeFileAsync;
 
     // mkdirAsync
-    t.is(mkdirAsyncFn.callCount, 2);
-    t.true(mkdirAsyncFn.args[0][0].includes(path.join(ruleScriptDir, results.name)));
-    t.true(mkdirAsyncFn.args[1][0].includes(path.join(ruleTestDir, results.name)));
+    t.is(mkdirpAsyncFn.callCount, 3);
+    t.true(mkdirpAsyncFn.args[0][0].includes(path.join(ruleScriptDir, results.name)));
+    t.true(mkdirpAsyncFn.args[1][0].includes(path.join(ruleDocDir, ''))); // so it uses the right separator
+    t.true(mkdirpAsyncFn.args[2][0].includes(path.join(ruleTestDir, results.name)));
+
 
     // writeFileAsync
     t.is(writeFileAsyncFn.callCount, 4);
@@ -109,39 +111,6 @@ test.serial(`if core, 'generate' should call to write script, documentation, tes
 
     sandbox.restore();
 });
-
-// test.serial(`if not core, 'generate' should only call to write script`, async (t) => {
-//     const sandbox = sinon.sandbox;
-//     const results = {
-//         category: 'PWAs',
-//         description: 'An important new rule',
-//         elementType: '',
-//         extension: 'ts',
-//         isCore: false,
-//         name: newRuleName,
-//         useCase: 'request'
-//     };
-
-//     sandbox.stub(stubPromisifiedMethodObject, 'readFileAsync').resolves('');
-//     sandbox.stub(inquirer, 'prompt').resolves(results);
-
-//     t.context.sandbox = sandbox;
-//     t.context.readFileAsync = stubPromisifiedMethodObject.readFileAsync;
-//     await rule.generate();
-
-//     const mkdirAsyncFn = t.context.mkdirAsync;
-//     const writeFileAsyncFn = t.context.writeFileAsync;
-
-//     // mkdirAsync
-//     t.is(mkdirAsyncFn.callCount, 1);
-//     t.true(mkdirAsyncFn.args[0][0].includes(path.join(ruleScriptDir, results.name)));
-
-//     // writeFileAsync
-//     t.is(writeFileAsyncFn.callCount, 1);
-//     t.true(writeFileAsyncFn.args[0][0].includes(path.join(ruleScriptDir, results.name, `${results.name}.${results.extension}`)));
-
-//     t.context.sandbox.restore();
-// });
 
 test.serial(`The right script template should be used in 'generate'`, async (t) => {
     const sandbox = sinon.sandbox.create();
@@ -161,7 +130,7 @@ test.serial(`The right script template should be used in 'generate'`, async (t) 
     sandbox.stub(inquirer, 'prompt').resolves(results);
     sandbox.stub(stubPromisifiedMethodObject, 'readFileAsync').resolves(scriptTemplate);
 
-    await rule.newRule();
+    await rule.newRule(actions);
 
     const expectedContent = await readFileAsync(expectedScriptDir);
     const actualContent = t.context.writeFileAsync.args[0][1];
@@ -189,7 +158,7 @@ test.serial(`Description contains quotes`, async (t) => {
     sandbox.stub(inquirer, 'prompt').resolves(results);
     sandbox.stub(stubPromisifiedMethodObject, 'readFileAsync').resolves(scriptTemplate);
 
-    await rule.newRule();
+    await rule.newRule(actions);
 
     const expectedContent = await readFileAsync(expectedScriptHasQuotesDir);
     const actualContent = t.context.writeFileAsync.args[0][1];
@@ -198,31 +167,6 @@ test.serial(`Description contains quotes`, async (t) => {
 
     sandbox.restore();
 });
-
-// test.serial(`Throw an error if it's a core rule but 'js' is used as an extension when calling 'generate'`, async (t) => {
-//     const sandbox = sinon.sandbox;
-//     const results = {
-//         category: 'PWAs',
-//         description: 'An important new rule',
-//         elementType: '',
-//         extension: 'js',
-//         isCore: true,
-//         name: newRuleName,
-//         useCase: 'request'
-//     };
-
-//     sandbox.stub(stubPromisifiedMethodObject, 'readFileAsync').resolves('');
-//     sandbox.stub(inquirer, 'prompt').resolves(results);
-
-//     t.context.sandbox = sandbox;
-//     t.context.readFileAsync = stubPromisifiedMethodObject.readFileAsync;
-
-//     const error = await t.throws(rule.generate());
-
-//     t.is(error.message, `The core rule can't be a 'js' file.`);
-
-//     t.context.sandbox.restore();
-// });
 
 test.serial(`Throw an error if a new rule already exists when calling 'generate'`, async (t) => {
     const sandbox = sinon.sandbox.create();
@@ -239,54 +183,15 @@ test.serial(`Throw an error if a new rule already exists when calling 'generate'
     sandbox.stub(stubPromisifiedMethodObject, 'readFileAsync').resolves('');
     sandbox.stub(inquirer, 'prompt').resolves(results);
 
-    const error = await t.throws(rule.newRule());
+    const error = await t.throws(rule.newRule(actions));
 
     t.is(error.message, `This rule already exists!`);
 
     sandbox.restore();
 });
 
-test.serial(`'remove' should call to remove the correct rule`, async (t) => {
-    const sandbox = sinon.sandbox.create();
-    const results = { name: existingRuleName };
+test.serial('if version is not an option, it should return false', async (t) => {
+    const result = await rule.newRule(({}) as CLIOptions);
 
-    sandbox.stub(stubPromisifiedMethodObject, 'readFileAsync').resolves('');
-    sandbox.stub(inquirer, 'prompt').resolves(results);
-
-    await rule.removeRule();
-
-    const rimrafAsyncFn = t.context.rimrafAsync;
-    const writeFileAsyncFn = t.context.writeFileAsync;
-
-    // rimrafAsync
-    t.is(rimrafAsyncFn.callCount, 4);
-    // Delete script
-    t.true(rimrafAsyncFn.args[0][0].includes(path.join(ruleScriptDir, normalizedExistingRuleName)));
-    // Delete documentation
-    t.true(rimrafAsyncFn.args[1][0].includes(path.join(ruleDocDir, `${normalizedExistingRuleName}.md`)));
-    // Delete test
-    t.true(rimrafAsyncFn.args[2][0].includes(path.join(ruleTestDir, normalizedExistingRuleName)));
-    // Delete dist script
-    t.true(rimrafAsyncFn.args[3][0].includes(path.join(ruleDistScriptDir, normalizedExistingRuleName)));
-
-    // writeFileAsync
-    t.is(writeFileAsyncFn.callCount, 1);
-    // UPdate index page
-    t.true(writeFileAsyncFn.args[0][0].includes(path.join(ruleDocDir, 'index.md')));
-
-    sandbox.restore();
-});
-
-test.serial(`Throw an error if a rule doesn't exist when calling 'remove'`, async (t) => {
-    const sandbox = sinon.sandbox.create();
-    const results = { name: newRuleName };
-
-    sandbox.stub(stubPromisifiedMethodObject, 'readFileAsync').resolves('');
-    sandbox.stub(inquirer, 'prompt').resolves(results);
-
-    const error = await t.throws(rule.removeRule());
-
-    t.is(error.message, `This rule doesn't exist!`);
-
-    sandbox.restore();
+    t.false(result);
 });
