+++ tests/lib/cli/rules/fixtures/new-quotes.txt
