+++ tests/lib/cli/rules/fixtures/new.txt
